remove Clone() in favor of increasing the reference counter.
The thing is, when ever we use clone() it increases memory allocation
and cpu use in favor of simplified coding. But i've always been puzzled
why the reference counter wasn't implemented. Its not hard, the real
problem with it is that we don't want objects dangling around. So we need
to add some sort of system that allows us to track the objects and check
if they are dangling free... lol