Change mailLabel table structure
The old table structure (and the accompanying code) suffers from some heavy structural problems. Further research in the client code made it apparent that labelID is not a globally unique id, it is instead unique only to the owner. Also, its not an actual ID; it is rather an index into a bitset, then to be used by a labelMask. It is as this point not clear why the EVE devs decided to go with such a ridiculously 'efficient' (and hard to build, maintain or expand) implementation, given that traffic isn't of much concern to them comparing to other subsystems

Signed-off-by: Stefan Schake <caytchen@gmail.com>