Revert "Added '%' signs to PRI/SCN defines as they were used fro printf"
This is not the correct solution; the proper fix is to add the '%'
sign wherever the defines are used (see the documentation of
inttypes.h at opengroup.org for details), which is what I am going to
do in the next commit.

Nevertheless, nice one spotting this, oza - it's my fault in the first
place.

This reverts commit 03e01ce0e180b4d8b85f8c69310766aabd7f02ec.