Fixed docker permissions issue
Due to a limitation of docker-compose and podman-compose, volumes are always mounted as root.

Our image doesn't have in-container root permissions, so caches fail to work due to permission failures.

Running as root in this container is not a problem since eve-server is the only process running. There is no security issue here.

This commit removes the unnecessary user and group from the dockerfile.