Enable GDB and interactive mode in Docker
This commit allows the use of server interactive mode by simply `docker attach <container>` to the server container.

It also allows the use of GDB to run the server by passing RUN_WITH_GDB=TRUE to the program.

This will allow the setting of breakpoints and collecting of backtraces easily.