Fixed alliance notifications
This commit fixes previous issues with alliance notifications due to Python Long type conversion.

We also rework OnAllianceChanged and OnAllianceMemberChanged.

I have tested and now a player can create an Alliance and also apply to join alliance (incl. rejection) and notifications are now sent so client does not need to be manually refreshed.