Save coordinates when logging out: In WarpOut(), get the current precise coordinates of the ship, and then call m_ship->SaveShip() to save them to the database.
Database saving mechanism: SaveShip() calls ItemDB::SaveItem(), which saves the coordinates (x, y, z) of the ship to the corresponding fields of the entity table.
Restore on next login: When the character logs in next time, the system will read the saved coordinates of the ship from the database instead of using the default point or random point.